Move all "should deallocate" tests to SignalLifetimeSpec.swift
